Struct replicapoolupdater1_beta1::RollingUpdatePolicy

pub struct RollingUpdatePolicy {
    pub sleep_after_instance_restart_sec: Option<i32>,
    pub instance_startup_timeout_sec: Option<i32>,
    pub max_num_concurrent_instances: Option<i32>,
    pub min_instance_update_time_sec: Option<i32>,
    pub auto_pause_after_instances: Option<i32>,
    pub max_num_failed_instances: Option<i32>,
}

Parameters of the update process.

This type is not used in any activity, and only used as part of another schema.

Fields

sleep_after_instance_restart_sec

Time period after the instance has been restarted but before marking the update of this instance as done. This field is deprecated and ignored by Rolling Updater.

instance_startup_timeout_sec

Maximum amount of time we will wait after finishing all steps until we receive HEALTHY state for instance. If this deadline is exceeded instance update is considered as failed.

max_num_concurrent_instances

Maximum number of instances that can be updated simultaneously (concurrently). An update of an instance starts when the instance is about to be restarted and finishes after the instance has been restarted and the sleep period (defined by sleepAfterInstanceRestartSec) has passed.

min_instance_update_time_sec

Specifies minimum amount of time we will spend on updating single instance, measuring at the start of the first update action (e.g. Recreate call on Instance Group Manager or Stop call on Instance resource). If actual instance update takes less time we will simply sleep before proceeding with next instance.

auto_pause_after_instances

Number of instances updated before the update gets automatically paused.

max_num_failed_instances

Maximum number of instance updates that can fail without failing the group update. Instance update is considered failed if any of it's update actions (e.g. Stop call on Instance resource in Rolling Reboot) failed with permanent failure, or if after finishing all update actions this instance is in UNHEALTHY state.